removing extra file for wrapper code
