--- cartofacade/carto_facade_wrapper_test.go
@@ -1,409 +1,0 @@
-//nolint:dupl
-package cartofacade
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"os"
-	"sync"
-	"testing"
-	"time"
-
-	"go.viam.com/rdk/pointcloud"
-	"go.viam.com/test"
-	"go.viam.com/utils/artifact"
-)
-
-func TestInitialize(t *testing.T) {
-	lib, err := NewLib(1, 1)
-	test.That(t, err, test.ShouldBeNil)
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-
-	t.Run("test Initialize - successful case", func(t *testing.T) {
-		err = cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-		test.That(t, err, test.ShouldBeNil)
-	})
-
-	lib.Terminate()
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestStart(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.StartFunc = func() error {
-		return nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing Start", func(t *testing.T) {
-		// success case
-		err = cartoFacade.Start(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-
-		carto.StartFunc = func() error {
-			return errors.New("test error 1")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		err = cartoFacade.Start(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 1"))
-
-		carto.StartFunc = func() error {
-			time.Sleep(1 * time.Second)
-			return nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		err = cartoFacade.Start(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestStop(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.StopFunc = func() error {
-		return nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing Stop", func(t *testing.T) {
-		// success case
-		err = cartoFacade.Stop(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-
-		carto.StopFunc = func() error {
-			return errors.New("test error 2")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		err = cartoFacade.Stop(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 2"))
-
-		carto.StopFunc = func() error {
-			time.Sleep(1 * time.Second)
-			return nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		err = cartoFacade.Stop(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestTerminate(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.TerminateFunc = func() error {
-		return nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing Terminate", func(t *testing.T) {
-		// success case
-		err = cartoFacade.Terminate(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-
-		carto.TerminateFunc = func() error {
-			return errors.New("test error 3")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		err = cartoFacade.Terminate(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 3"))
-
-		carto.TerminateFunc = func() error {
-			time.Sleep(1 * time.Second)
-			return nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		err = cartoFacade.Terminate(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestAddSensorReading(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.AddSensorReadingFunc = func(name string, reading []byte, time time.Time) error {
-		return nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing AddSensorReading", func(t *testing.T) {
-		timestamp := time.Date(2021, 8, 15, 14, 30, 45, 100, time.Local)
-
-		// read PCD
-		file, err := os.Open(artifact.MustPath("viam-cartographer/mock_lidar/0.pcd"))
-		test.That(t, err, test.ShouldBeNil)
-		buf := new(bytes.Buffer)
-		pc, err := pointcloud.ReadPCD(file)
-		test.That(t, err, test.ShouldBeNil)
-		err = pointcloud.ToPCD(pc, buf, 0)
-		test.That(t, err, test.ShouldBeNil)
-
-		// success case
-		err = cartoFacade.AddSensorReading(cancelCtx, 5*time.Second, "mysensor", buf.Bytes(), timestamp)
-		test.That(t, err, test.ShouldBeNil)
-
-		carto.AddSensorReadingFunc = func(name string, reading []byte, time time.Time) error {
-			return errors.New("test error 4")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		err = cartoFacade.AddSensorReading(cancelCtx, 5*time.Second, "mysensor", buf.Bytes(), timestamp)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 4"))
-
-		carto.AddSensorReadingFunc = func(name string, reading []byte, timestamp time.Time) error {
-			time.Sleep(1 * time.Second)
-			return nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		err = cartoFacade.AddSensorReading(cancelCtx, 1*time.Millisecond, "mysensor", buf.Bytes(), timestamp)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestGetPosition(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.GetPositionFunc = func() (GetPosition, error) {
-		pos := GetPosition{X: 1, Y: 2, Z: 3}
-		return pos, nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing GetPosition", func(t *testing.T) {
-		// success case
-		pos, err := cartoFacade.GetPosition(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, pos.X, test.ShouldEqual, 1)
-		test.That(t, pos.Y, test.ShouldEqual, 2)
-		test.That(t, pos.Z, test.ShouldEqual, 3)
-
-		carto.GetPositionFunc = func() (GetPosition, error) {
-			return GetPosition{}, errors.New("test error 5")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		_, err = cartoFacade.GetPosition(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 5"))
-
-		carto.GetPositionFunc = func() (GetPosition, error) {
-			time.Sleep(1 * time.Second)
-			return GetPosition{}, nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		_, err = cartoFacade.GetPosition(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestGetInternalState(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.GetInternalStateFunc = func() ([]byte, error) {
-		internalState := []byte("hello!")
-		return internalState, nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing GetInternalState", func(t *testing.T) {
-		// success case
-		internalState, err := cartoFacade.GetInternalState(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, internalState, test.ShouldResemble, []byte("hello!"))
-
-		carto.GetInternalStateFunc = func() ([]byte, error) {
-			return []byte{}, errors.New("test error 6")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		_, err = cartoFacade.GetInternalState(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 6"))
-
-		carto.GetInternalStateFunc = func() ([]byte, error) {
-			time.Sleep(1 * time.Second)
-			return []byte{}, nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		_, err = cartoFacade.GetInternalState(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
-
-func TestGetPointCloudMap(t *testing.T) {
-	lib := CartoLibMock{}
-
-	cancelCtx, cancelFunc := context.WithCancel(context.Background())
-	activeBackgroundWorkers := sync.WaitGroup{}
-
-	cfg, dir, err := GetTestConfig("mysensor")
-	algoCfg := GetTestAlgoConfig()
-	test.That(t, err, test.ShouldBeNil)
-	defer os.RemoveAll(dir)
-
-	cartoFacade := New(&lib, cfg, algoCfg)
-	carto := CartoMock{}
-	carto.GetPointCloudMapFunc = func() ([]byte, error) {
-		internalState := []byte("hello!")
-		return internalState, nil
-	}
-	cartoFacade.carto = &carto
-	cartoFacade.Initialize(cancelCtx, 5*time.Second, &activeBackgroundWorkers)
-
-	t.Run("testing GetPointCloudMap", func(t *testing.T) {
-		// success case
-		internalState, err := cartoFacade.GetPointCloudMap(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeNil)
-		test.That(t, internalState, test.ShouldResemble, []byte("hello!"))
-
-		carto.GetPointCloudMapFunc = func() ([]byte, error) {
-			return []byte{}, errors.New("test error 7")
-		}
-		cartoFacade.carto = &carto
-
-		// returns error
-		_, err = cartoFacade.GetPointCloudMap(cancelCtx, 5*time.Second)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("test error 7"))
-
-		carto.GetPointCloudMapFunc = func() ([]byte, error) {
-			time.Sleep(1 * time.Second)
-			return []byte{}, nil
-		}
-		cartoFacade.carto = &carto
-
-		// times out
-		_, err = cartoFacade.GetPointCloudMap(cancelCtx, 1*time.Millisecond)
-		test.That(t, err, test.ShouldBeError)
-		test.That(t, err, test.ShouldResemble, errors.New("timeout has occurred while trying to read request from cartofacade"))
-	})
-
-	cancelFunc()
-	activeBackgroundWorkers.Wait()
-}
